+++ OpenHD/ohd_common/CMakeLists.txt
@@ -51,8 +51,8 @@ set(sources
     "openhd-platform-json.hpp"
     "openhd-profile.hpp"
     "openhd-profile-json.hpp"
-    "openhd-settings.hpp"
-    "openhd-settings2.hpp"
+    "openhd-settings-directories.hpp"
+    "openhd-settings-persistent.hpp"
     "openhd-util.hpp"
     "openhd-util-filesystem.hpp"
     "veye-helper.hpp"
