+++ OpenHD/ohd_common/openhd-settings-persistent.hpp
@@ -2,8 +2,8 @@
 // Created by consti10 on 18.07.22.
 //
 
-#ifndef OPENHD_OPENHD_OHD_COMMON_OPENHD_SETTINGS2_HPP_
-#define OPENHD_OPENHD_OHD_COMMON_OPENHD_SETTINGS2_HPP_
+#ifndef OPENHD_OPENHD_OHD_COMMON_OPENHD_SETTINGS_PERSISTENT_HPP_
+#define OPENHD_OPENHD_OHD_COMMON_OPENHD_SETTINGS_PERSISTENT_HPP_
 
 #include <utility>
 
@@ -143,4 +143,4 @@ class PersistentSettings{
 
 }
 
-#endif //OPENHD_OPENHD_OHD_COMMON_OPENHD_SETTINGS2_HPP_
+#endif  // OPENHD_OPENHD_OHD_COMMON_OPENHD_SETTINGS_PERSISTENT_HPP_
