+++ OpenHD/ohd_telemetry/src/AirTelemetrySettings.h
@@ -5,9 +5,10 @@
 #ifndef OPENHD_OPENHD_OHD_TELEMETRY_SRC_AIRTELEMETRYSETTINGS_H_
 #define OPENHD_OPENHD_OHD_TELEMETRY_SRC_AIRTELEMETRYSETTINGS_H_
 
-#include "openhd-settings.hpp"
-#include "openhd-settings2.hpp"
 #include <map>
+
+#include "openhd-settings-directories.hpp"
+#include "openhd-settings-persistent.hpp"
 
 // Settings for telemetry, only valid on an air pi (since only on the air pi we connect the FC)
 // Note that we do not have any telemetry settings r.n for the ground (since forwarding is a task ov ohd_interface, not ohd_telemetry)
